Rename example project
