+++ CircularBuffer.Example/app.config
